Syncing after downloading tar.bz2
[git-p4: depot-paths = "//depot/openwatcom/": change = 26500]

+++ bld/gui/win/h/mem.h
@@ -31,5 +31,5 @@
 
 
 void MemFree( void *ptr );
-void *MemAlloc( size_t size );
-void *MemReAlloc( void *ptr, size_t size );
+void *MemAlloc( unsigned size );
+void *MemReAlloc( void *ptr, unsigned size );
